Add Mathematica-based QR-code silkscreen creation
EagleQR.nb will generate a QR code library with a schematic part as well
as silk screens.  Positive and negative versions are provided for use with
light silk on dark solder mask or the other way around.

Could be ported to python if a suitable qr code thing is found, but I'm
lazy.